Now the images are placed correctly, since they're static we don't need to update uniforms, just create the original vbo with the correct corrdinates and that's it, no updates

+++ include/rGlobal.h
@@ -5,8 +5,8 @@
 
 typedef struct rGlobal {
 	const char* app_name;
-	unsigned int window_width;
-	unsigned int window_height;
+	float window_width;
+	float window_height;
 	unsigned int slide_count;
 	TTF_Font* font; // for now only one global font
 } rGlobal;
